requirements.txt
numpy==1.17.2
pytorch==1.3.0+cu92
transforms3d==0.3.1
h5py==2.9.0
ninja==1.9.0.post1
open3d==0.7.0.0
tqdm==4.35.0
tensorboardX=1.8
scipy==1.3.1
sklearn==0.21.3